creation of middleware directory, specifically the piece that provides the react html template
