+++ src/server/utils/getEnv.ts
@@ -1,0 +1,9 @@
+export const getEnv = () => {
+  const env = process.env.NODE_ENV;
+
+  if (env === undefined) {
+    return "local";
+  }
+
+  return env;
+};
